[devtools] refactor segment explorer styles
